Fix bug preventing `asr` prop from being reactive
After we construct the reactive `props` object, we merge it with the default options. However, if there was a `props` object on the `defaultOptions` object, it would break the reactivity by merging the reactive props object with the non-reactive default props object. Instead, separate the default props off to prevent that.